backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
azure-cognitiveservices-speech==1.47.0
azure-cognitiveservices-vision-computervision==0.9.1
azure-common==1.1.28
azure-core==1.38.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
easyocr==1.7.0
edge-tts==6.1.19
Flask==3.0.3
Flask-Cors==4.0.1
fastapi==0.109.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
isodate==0.7.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.7
langchain-openai==1.1.7
langsmith==0.6.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
msrest==0.7.1
multidict==6.7.0
networkx==3.4.2
ninja==1.13.0
numpy==1.26.3
oauthlib==3.3.1
openai==2.15.0
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
orjson==3.11.5
packaging==25.0
pandas==2.2.3
pillow==12.0.0
platformdirs==4.5.1
propcache==0.4.1
protobuf==6.33.4
pyclipper==1.4.0
pyproj==3.6.1
pydantic==2.12.5
pydantic_core==2.41.5
python-bidi==0.6.7
python-dotenv==1.0.1
python-multipart==0.0.6
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
sentry-sdk==2.50.0
setuptools==70.2.0
shapely==2.1.2
smmap==5.0.2
sniffio==1.3.1
starlette==0.35.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2026.1.14
tiktoken==0.12.0
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.27.0
wandb==0.24.0
watchfiles==1.1.1
websockets==16.0
yarl==1.22.0
zstandard==0.25.0